UI: Extract sensitive input into it's own component
The main benefit is that we reduce the duplication in styling and icons
