+++ monkey/monkey_island/cc/ui/src/components/ui-components/SensitiveTextareaInput.tsx
@@ -1,35 +1,29 @@
 import React, {useState} from 'react';
-import {InputGroup, FormControl} from 'react-bootstrap';
+import {FormControl} from 'react-bootstrap';
+import SensitiveInput from './SensitiveInput';
 
 function SensitiveTextareaInput(props){
-    const [hidden, setHidden] = useState(false);
+  const [hidden, setHidden] = useState(false);
 
-    const toggleShow = () =>{
-        setHidden(! hidden);
-    }
+  const onChange = (value) => {
+    return props.onChange(value === '' ? props.options.emptyValue : value);
+  }
 
-    const onChange = (value) => {
-        return props.onChange(value === '' ? props.options.emptyValue : value);
-    }
+  let inputComponent = (
+    <FormControl
+      as="textarea"
+      rows={5}
+      className={hidden ? '' : 'password'}
+      value={props.value || ''}
+      onChange={(event) => onChange(event.target.value)}
+    />
+  )
 
-    return (
-    <div>
-      <InputGroup>
-        <FormControl
-            as="textarea"
-            rows={5}
-            className={hidden ? '' : 'password'}
-            value={props.value || ''}
-            onChange={(event) => onChange(event.target.value)}
-        />
-        <InputGroup.Append>
-          <InputGroup.Text onClick={toggleShow} >
-            <i className={hidden ? 'fas fa-eye-slash' : 'fas fa-eye'}></i>
-          </InputGroup.Text>
-        </InputGroup.Append>
-      </InputGroup>
-    </div>
-    );
+  return (
+    <SensitiveInput inputComponent={inputComponent}
+                    toggle={() => setHidden(!hidden)}
+                    hidden={hidden}/>
+  );
 }
 
 export default SensitiveTextareaInput;
